fix journeymap absence check
