+++ mod/common/src/main/java/gjum/minecraft/mapsync/common/integration/JourneyMapHelper.java
@@ -1,38 +1,34 @@
 package gjum.minecraft.mapsync.common.integration;
 
-import gjum.minecraft.mapsync.common.data.*;
+import gjum.minecraft.mapsync.common.data.ChunkTile;
 import journeymap.client.JourneymapClient;
 import journeymap.client.io.FileHandler;
-import journeymap.client.model.*;
+import journeymap.client.model.MapType;
+import journeymap.client.model.RegionCoord;
 import journeymap.common.nbt.RegionData;
 import journeymap.common.nbt.RegionDataStorageHandler;
-import net.minecraft.core.BlockPos;
-import net.minecraft.world.level.biome.Biome;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.chunk.LevelChunk;
-
-import java.util.List;
 
 import static gjum.minecraft.mapsync.common.Utils.mc;
 
 public class JourneyMapHelper {
-	public static boolean isJourneyMapNotAvailable() {
+	public static boolean isJourneyMapNotAvailable;
+
+	static {
 		try {
 			Class.forName("journeymap.client.JourneymapClient");
-			return false;
+			isJourneyMapNotAvailable = false;
 		} catch (NoClassDefFoundError | ClassNotFoundException ignored) {
-			return true;
+			isJourneyMapNotAvailable = true;
 		}
 	}
 
 	public static boolean isMapping() {
-		if (isJourneyMapNotAvailable()) return false;
+		if (isJourneyMapNotAvailable) return false;
 		return JourneymapClient.getInstance().isMapping();
 	}
 
 	public static boolean updateWithChunkTile(ChunkTile chunkTile) {
-		if (isJourneyMapNotAvailable()) return false;
+		if (isJourneyMapNotAvailable) return false;
 		if (!JourneymapClient.getInstance().isMapping()) return false; // BaseMapTask does this
 
 		var renderController = JourneymapClient.getInstance().getChunkRenderController();
@@ -62,78 +58,5 @@ public class JourneyMapHelper {
 		if (!renderedTopo) System.out.println("Failed rendering topo at " + chunkTile.chunkPos());
 
 		return renderedDay && renderedBiome && renderedTopo;
-	}
-
-	private static class TileChunkMD extends NBTChunkMD {
-		private final ChunkTile chunkTile;
-
-		public TileChunkMD(ChunkTile chunkTile) {
-			super(new LevelChunk(mc.level, chunkTile.chunkPos()),
-					chunkTile.chunkPos(),
-					null, // all accessing methods are overridden
-					MapType.day(chunkTile.dimension()) // just has to not be `underground`
-			);
-			this.chunkTile = chunkTile;
-		}
-
-		@Override
-		public boolean hasChunk() {
-			return true;
-		}
-
-		private BlockColumn getCol(int x, int z) {
-			int xic = x & 0xf;
-			int zic = z & 0xf;
-			return chunkTile.columns()[xic + zic * 16];
-		}
-
-		private BlockColumn getCol(BlockPos pos) {
-			return getCol(pos.getX(), pos.getZ());
-		}
-
-		@Override
-		public BlockState getBlockState(BlockPos pos) {
-			var layers = getCol(pos.getX(), pos.getZ()).layers();
-			BlockInfo prevLayer = null;
-			// note that layers are ordered top-down
-			for (BlockInfo layer : layers) {
-				if (layer.y() == pos.getY()) {
-					return layer.state();
-				}
-				if (layer.y() < pos.getY()) {
-					// top of layer is below pos, so pos is inside prevLayer
-					if (prevLayer == null) return null; // first layer is already below pos
-					return prevLayer.state();
-				}
-				prevLayer = layer;
-			}
-			if (layers.isEmpty()) return Blocks.AIR.defaultBlockState();
-			return getLast(layers).state();
-		}
-
-		@Override
-		public Integer getGetLightValue(BlockPos pos) {
-			return getCol(pos.getX(), pos.getZ()).light();
-		}
-
-		@Override
-		public Integer getTopY(BlockPos pos) {
-			return getCol(pos.getX(), pos.getZ()).layers().get(0).y();
-		}
-
-		@Override
-		public int getHeight(BlockPos pos) {
-			return this.getTopY(pos);
-		}
-
-		@Override
-		public Biome getBiome(BlockPos pos) {
-			return getCol(pos).biome();
-		}
-	}
-
-	private static <T> T getLast(List<T> l) {
-		if (l.isEmpty()) throw new Error("Empty list");
-		return l.get(l.size() - 1);
 	}
 }
